testing/requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiohttp-swagger==1.0.15
ansible==4.3.0
ansible-core==2.11.3
anyio==3.3.0
arrow==1.1.1
async-lru==1.0.2
async-timeout==3.0.1
attrs==21.2.0
backports.ssl==0.0.9
binaryornot==0.4.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
cookiecutter==1.7.3
cryptography==3.4.7
dill==0.3.4
distro==1.5.0
dnspython==2.1.0
genie==21.6
genie.libs.clean==21.6
genie.libs.conf==21.6
genie.libs.filetransferutils==21.6
genie.libs.health==21.6.1
genie.libs.ops==21.6
genie.libs.parser==21.6
genie.libs.robot==21.6
genie.libs.sdk==21.6.1
genie.telemetry==21.6
genie.trafficgen==21.6
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.13.6
httpx==0.18.2
hyperframe==5.2.0
idna==3.2
IxNetwork==9.0.1915.16
ixnetwork-restpy==1.0.67
Jinja2==2.11.3
jinja2-time==0.2.0
jsonpickle==2.0.0
junit-xml==1.9
MarkupSafe==1.1.1
multidict==5.1.0
nested-lookup==0.2.22
netaddr==0.8.0
packaging==21.0
pathspec==0.9.0
poyo==0.5.0
prettytable==2.1.0
psutil==5.8.0
pyats==21.6
pyats.aereport==21.6
pyats.aetest==21.6
pyats.async==21.6
pyats.connections==21.6
pyats.contrib==21.6
pyats.datastructures==21.6
pyats.easypy==21.6
pyats.kleenex==21.6.1
pyats.log==21.6
pyats.reporter==21.6
pyats.results==21.6
pyats.robot==21.6
pyats.tcl==21.6
pyats.topology==21.6
pyats.utils==21.6
pycparser==2.20
pyftpdlib==1.5.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-engineio==3.14.2
python-slugify==5.0.2
python-socketio==4.6.1
PyYAML==5.4.1
pyyaml-include==1.2.post2
requests==2.26.0
requests-toolbelt==0.9.1
resolvelib==0.5.4
rfc3986==1.5.0
robotframework==4.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sniffio==1.2.0
text-unidecode==1.3
tftpy==0.8.0
tqdm==4.61.2
typing-extensions==3.10.0.0
unicon==21.6
unicon.plugins==21.6
urllib3==1.26.6
virl2-client==2.2.1.post1
wcwidth==0.2.5
websocket-client==1.1.0
xlrd==1.2.0
XlsxWriter==1.4.4
xlwt==1.3.0
xmltodict==0.12.0
yamllint==1.26.1
yarl==1.6.3
